+++ src/main/kotlin/com/conceptic/firefly/app/screen/support/MouseActionsPublisher.kt
@@ -6,15 +6,15 @@ import com.conceptic.firefly.support.Publisher
 interface MouseActionsSubscriber : MouseListener
 
 class MouseActionsPublisher : Publisher<MouseActionsSubscriber>(), MouseListener {
-    override fun onClicked(x: Int, y: Int) {
+    override fun onClicked(x: Float, y: Float) {
         notify { onClicked(x, y) }
     }
 
-    override fun onPressed(x: Int, y: Int) {
+    override fun onPressed(x: Float, y: Float) {
         notify { onPressed(x, y) }
     }
 
-    override fun onMoved(x: Int, y: Int) {
+    override fun onMoved(x: Float, y: Float) {
         notify { onMoved(x, y) }
     }
 }
